chore: Move EnvironmentAwareIntegrationTest out of Integration test namespace and make it abstract
