+++ tests/EnvironmentAwareIntegrationTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Laudis\Neo4j\Tests\Integration;
+namespace Laudis\Neo4j\Tests;
 
 use function is_string;
 
@@ -20,7 +20,7 @@ use Laudis\Neo4j\Basic\Session;
 use Laudis\Neo4j\Common\Uri;
 use PHPUnit\Framework\TestCase;
 
-class EnvironmentAwareIntegrationTest extends TestCase
+abstract class EnvironmentAwareIntegrationTest extends TestCase
 {
     protected static Session $session;
     protected static Driver $driver;
